feat(contracts): add safe factory on genesis (#437)
Porto AA contracts rely on the safe singleton deployer to be deployed on
the canonical address, to ensure that all contracts on all chains have
the same address.

So this factory needs to be added to the next testnet, for porto infra
to work out of the box with Tempo.

---------

Co-authored-by: Jennifer <jenpaff0@gmail.com>